fix: admin_books_route, admin/books_route 통합, 안 쓰는 current_user 제거
